Refactor: Move labeling components to step06 and update import paths
