+++ src/training/early_stage_optimization.py
@@ -21,7 +21,7 @@ try:
 except ImportError:
     OPTUNA_AVAILABLE = False
 try:
-    from .steps.step4_analyst_labeling_feature_engineering_components.regime_specific_triple_barrier_optimizer import RegimeSpecificTripleBarrierOptimizer, create_regime_specific_triple_barrier_optimizer
+    from .steps.step06_labeling_components.regime_specific_triple_barrier_optimizer import RegimeSpecificTripleBarrierOptimizer, create_regime_specific_triple_barrier_optimizer
     REGIME_OPTIMIZER_AVAILABLE = True
 except ImportError:
     REGIME_OPTIMIZER_AVAILABLE = False
