+++ src/training/steps/model_training/step04_5_triple_barrier_method.py
@@ -51,7 +51,7 @@ class TripleBarrierMethodStep:
         """Initialize triple barrier method components."""
         self.logger.info('🔧 Initializing triple barrier method components...')
         try:
-            from .step4_analyst_labeling_feature_engineering_components.optimized_triple_barrier_labeling import (
+            from .step06_labeling_components.optimized_triple_barrier_labeling import (
                 OptimizedTripleBarrierLabeling
             )
             self.triple_barrier_labeler = OptimizedTripleBarrierLabeling()
