+++ src/training/steps/step06_labeling_components/__init__.py
@@ -1,11 +1,11 @@
 from __future__ import annotations
 
-# src/training/steps/step4_analyst_labeling_feature_engineering_components/__init__.py
+# src/training/steps/step06_labeling_components/__init__.py
 
-"""Step 4 Analyst Labeling and Feature Engineering Components.
+"""Step 06 Labeling Components.
 
-This module contains the components for triple barrier labeling and feature engineering
-used in the analyst training pipeline.
+This module contains the components for triple barrier labeling used in the training pipeline.
+Feature engineering has been moved to separate components.
 """
 
 from .fractional_triple_barrier_labeling import FractionalTripleBarrierLabeling
@@ -28,5 +28,5 @@ __all__ = [
 __version__ = "1.0.0"
 __author__ = "Ares Trading System"
 __description__ = (
-    "Optimized triple barrier labeling and feature engineering components for step 4"
+    "Optimized triple barrier labeling components for step 06"
 )
